+++ testing/conftest.py
@@ -1,3 +1,4 @@
+import os, pathlib, pytest
 from pytest import fixture
 
 
@@ -10,3 +11,19 @@ def pytest_addoption(parser):
 @fixture()
 def device(request):
     return request.config.getoption("--device")
+
+
+@pytest.hookimpl(tryfirst=True, hookwrapper=True)
+def pytest_runtest_makereport(item, call):
+    # execute all other hooks to obtain the report object
+    outcome = yield
+    rep = outcome.get_result()
+
+    # we only look at actual failing test calls, not setup/teardown
+    if rep.when == "call" and rep.failed:
+        log_filename = os.path.join(
+            pathlib.Path(__file__).parent.absolute(), "failures.log"
+        )
+        mode = "a" if os.path.exists(log_filename) else "w"
+        with open(log_filename, mode) as f:
+            f.write(rep.longreprtext + "\n")